\name{clean}
\alias{clean}
\title{clean}
\usage{
clean(corpus)
}
\description{
  # Convert text to lower case
    # Remove punctuation from document
      # Remove numbers from document
        # Remove common stop words in the english language
          # Converts word corpus to term document matrix
}
\arguments{
\item{corpus}{word corpus}}


