% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_punct.R
\name{clean_punct}
\alias{clean_punct}
\title{Clean punctuation and whitespace}
\usage{
clean_punct(text)
}
\arguments{
\item{text}{A character vector with text to format.}
}
\value{
A character vector with formatted text.
}
\description{
Adds space around punctuation, reduces multiple spaces to a single space,
removes spaces around apostrophes, and trims leading and trailing whitespace.
}
\examples{
clean_punct("Hello, world! 'This is an example.'")
}
