% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text}
\usage{
clean_text(
  text,
  remove_urls = TRUE,
  replace_amp = TRUE,
  replace_newline = TRUE,
  space_emoji = TRUE,
  remove_extra_spaces = TRUE,
  trim_spaces = TRUE
)
}
\arguments{
\item{text}{A character string containing the text to be cleaned.}

\item{remove_urls}{Logical, indicating whether to remove URLs.}

\item{replace_amp}{Logical, indicating whether to replace "&" with "and".}

\item{replace_newline}{Logical, indicating whether to replace newline characters with space.}

\item{space_emoji}{Logical, indicating whether to add space between text and emoji.}

\item{remove_extra_spaces}{Logical, indicating whether to remove extra spaces.}

\item{trim_spaces}{Logical, indicating whether to trim spaces at the beginning and end of the text.}
}
\value{
A cleaned text as a character string.
}
\description{
This function cleans text by removing URLs, replacing "&" with "and",
replacing newline characters with space, adding space between text and emoji, removing extra spaces,
and trimming spaces at the beginning and end of the text.
}
\examples{
clean_text("This is a test with a URL http://test.com &amp; some emojis 😊", remove_urls = TRUE, replace_amp = TRUE)
}
