% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_punctuation.R
\name{remove_punctuation}
\alias{remove_punctuation}
\title{Remove punctuation from text}
\usage{
remove_punctuation(text)
}
\arguments{
\item{text}{The input text from which to remove punctuation}
}
\value{
A character string without punctuation
}
\description{
Remove punctuation from text
}
\examples{
remove_punctuation("Hello, world!")
}
