% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_contract.R
\name{replace_contractions}
\alias{replace_contractions}
\title{Replace contractions and clean text}
\usage{
replace_contractions(
  text,
  custom_df = NULL,
  use_lexicon = TRUE,
  replace_trailing_s = TRUE,
  trim_whitespace = TRUE,
  replace_apostrophe = TRUE
)
}
\arguments{
\item{text}{A character vector containing the text to be cleaned.}

\item{custom_df}{An optional custom data frame containing contraction and expanded form mappings.}

\item{use_lexicon}{Logical, indicating whether to use the predefined contractions list from the 'lexicon' package.}

\item{replace_trailing_s}{Logical, indicating whether to replace trailing 's with 'is'.}

\item{trim_whitespace}{Logical, indicating whether to trim white space at the beginning and end of the text.}

\item{replace_apostrophe}{Logical, indicating whether to replace remaining apostrophes with a space.}
}
\value{
A character vector with cleaned text.
}
\description{
Replaces contractions with their expanded forms using either a predefined list from the 'lexicon' package or a custom list. Additionally, offers options to replace trailing 's with 'is', remove trailing white space, and replace remaining apostrophes with a space.
}
\examples{
replace_contractions("He's not going to the party because he doesn't want to.")
}
