% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardised_punctuation.R
\name{standardize_punctuation}
\alias{standardize_punctuation}
\title{Standardise Punctuation from Unicode to ASCII}
\usage{
standardize_punctuation(text)
}
\arguments{
\item{text}{A character vector containing the text to be cleaned.}
}
\value{
A character vector with standardised punctuation.
}
\description{
This function replaces specified Unicode punctuation characters with their ASCII equivalents in a given text.
}
\examples{
standardize_punctuation("‘This’ is “a” test—text")
}
